Add title truncation and optional line break after
Titles longer that 100 characters could cause wrapping. To resolve this
titles longer than 100 characters are now truncated to 97 characters and
are appended with an ellipsis. [closes #31]

Adding a Line break after titles can enhance the flow of the report and
make it easier to read. The option is --break-after-title. [closes #30]