Change handling of invalid|undefined report names
Modified handling of invalid or undefined report name to show a list of
report names along with a message requesting the user to select a name
from the list. This replaces the previous behavior which was to default
to the list-txt report. [closes #36]

This change was made along with the renaming of all the reports to
replace --report- with --issues-, which will allow for adding additional
report types such as those for GitHub reports. [closes #32]